Update README.md
Creating Data Products -- Programming Projects:
Week2P1.html:
This is a map created with Leaflet.
Week3P1All.html:
This the html presentation for Week3 Project 1. There are 59 slides in the deck, 11MB, it does take a few seconds to load. Be patient please. Graphs are interactive. Enjoy the read!